chore(server-config): Rename `tls_profile` into `network_encryption`
